+++ Backend/src/main/java/com/pz/ankietBud/model/subQuestion/Choice.java
@@ -21,13 +21,13 @@ public class Choice extends Question {
     public Choice() {
     }
 
-    public Choice(String content, ArrayList<Integer> votes, Integer sum, ArrayList<String> answers, Type type) {
+    public Choice(String content, ArrayList<Long> votes, Long sum, ArrayList<String> answers, Type type) {
         super(content, votes, sum);
         this.answers = answers;
         this.type = type;
     }
 
-    public Choice(Long id, String content, ArrayList<Integer> votes, Integer sum, ArrayList<String> answers, Type type) {
+    public Choice(Long id, String content, ArrayList<Long> votes, Long sum, ArrayList<String> answers, Type type) {
         super(id, content, votes, sum);
         this.answers = answers;
         this.type = type;
