+++ Backend/src/main/java/com/pz/ankietBud/model/subQuestion/Scale.java
@@ -17,7 +17,7 @@ public class Scale extends Question {
     private Integer value_min;
     private Integer value_max;
 
-    public Scale(String content, ArrayList<Integer> votes, Integer sum, String anchor_left, String anchor_right, String anchor_central, Integer value_min, Integer value_max) {
+    public Scale(String content, ArrayList<Long> votes, Long sum, String anchor_left, String anchor_right, String anchor_central, Integer value_min, Integer value_max) {
         super(content, votes, sum);
         this.anchor_left = anchor_left;
         this.anchor_right = anchor_right;
@@ -29,7 +29,7 @@ public class Scale extends Question {
     public Scale() {
     }
 
-    public Scale(Long id, String content, ArrayList<Integer> votes, Integer sum, String anchor_left, String anchor_right, String anchor_central, Integer value_min, Integer value_max) {
+    public Scale(Long id, String content, ArrayList<Long> votes, Long sum, String anchor_left, String anchor_right, String anchor_central, Integer value_min, Integer value_max) {
         super(id, content, votes, sum);
         this.anchor_left = anchor_left;
         this.anchor_right = anchor_right;
